+++ cpp/tests/community/ecg_test.cu
@@ -18,6 +18,7 @@
 
 TEST(ecg, success)
 {
+  // FIXME: verify that this is the karate dataset
   std::vector<int> off_h = {0,  16,  25,  35,  41,  44,  48,  52,  56,  61,  63, 66,
                             67, 69,  74,  76,  78,  80,  82,  84,  87,  89,  91, 93,
                             98, 101, 104, 106, 110, 113, 117, 121, 127, 139, 156};
@@ -63,11 +64,11 @@ TEST(ecg, success)
   cugraph::ext_raft::analyzeClustering_modularity(
     graph_csr, max + 1, result_v.data().get(), &modularity);
 
+  // 0.399 is 5% below the reference value returned in
+  // <cugraph>/python/utils/ECG_Golden.ipynb on the same dataset
   ASSERT_GT(modularity, 0.399);
 }
 
-//  This test currently fails... leaving it in since once louvain is fixed
-//   it should pass
 TEST(ecg, dolphin)
 {
   std::vector<int> off_h = {0,   6,   14,  18,  21,  22,  26,  32,  37,  43,  50,  55,  56,